ingestion_script/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-pinecone
pinecone-client[grpc]
pypdf
pymupdf
docx2txt
requests
beautifulsoup4
python-dotenv
llama-parse
unstructured[md]